Fix test and clean up code
