+++ src/main/java/org/opensearch/security/action/apitokens/ApiTokenIndexHandler.java
@@ -14,7 +14,6 @@ package org.opensearch.security.action.apitokens;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.function.Supplier;
 
 import com.google.common.collect.ImmutableMap;
 import org.apache.logging.log4j.LogManager;
@@ -27,7 +26,6 @@ import org.opensearch.action.search.SearchRequest;
 import org.opensearch.action.search.SearchResponse;
 import org.opensearch.client.Client;
 import org.opensearch.cluster.service.ClusterService;
-import org.opensearch.common.util.concurrent.ThreadContext;
 import org.opensearch.common.xcontent.XContentFactory;
 import org.opensearch.common.xcontent.XContentType;
 import org.opensearch.core.action.ActionListener;
@@ -42,7 +40,6 @@ import org.opensearch.index.reindex.DeleteByQueryAction;
 import org.opensearch.index.reindex.DeleteByQueryRequest;
 import org.opensearch.search.SearchHit;
 import org.opensearch.search.builder.SearchSourceBuilder;
-import org.opensearch.security.dlic.rest.support.Utils;
 import org.opensearch.security.support.ConfigConstants;
 
 import static org.opensearch.security.action.apitokens.ApiToken.NAME_FIELD;
@@ -59,116 +56,79 @@ public class ApiTokenIndexHandler {
     }
 
     public void indexTokenMetadata(ApiToken token) {
-        withSecurityContext(() -> {
-            try {
+        try {
 
-                XContentBuilder builder = XContentFactory.jsonBuilder();
-                String jsonString = token.toXContent(builder, ToXContent.EMPTY_PARAMS).toString();
+            XContentBuilder builder = XContentFactory.jsonBuilder();
+            String jsonString = token.toXContent(builder, ToXContent.EMPTY_PARAMS).toString();
 
-                IndexRequest request = new IndexRequest(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX).source(jsonString, XContentType.JSON);
+            IndexRequest request = new IndexRequest(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX).source(jsonString, XContentType.JSON);
 
-                ActionListener<IndexResponse> irListener = ActionListener.wrap(idxResponse -> {
-                    LOGGER.info("Created {} entry.", ConfigConstants.OPENSEARCH_API_TOKENS_INDEX);
-                }, (failResponse) -> {
-                    LOGGER.error(failResponse.getMessage());
-                    LOGGER.info("Failed to create {} entry.", ConfigConstants.OPENSEARCH_API_TOKENS_INDEX);
-                });
-                client.index(request, irListener);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        });
+            ActionListener<IndexResponse> irListener = ActionListener.wrap(idxResponse -> {
+                LOGGER.info("Created {} entry.", ConfigConstants.OPENSEARCH_API_TOKENS_INDEX);
+            }, (failResponse) -> {
+                LOGGER.error(failResponse.getMessage());
+                LOGGER.info("Failed to create {} entry.", ConfigConstants.OPENSEARCH_API_TOKENS_INDEX);
+            });
+            client.index(request, irListener);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
 
     }
 
     public void deleteToken(String name) throws ApiTokenException {
-        withSecurityContext(() -> {
-            DeleteByQueryRequest request = new DeleteByQueryRequest(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX).setQuery(
-                QueryBuilders.matchQuery(NAME_FIELD, name)
-            ).setRefresh(true);
+        DeleteByQueryRequest request = new DeleteByQueryRequest(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX).setQuery(
+            QueryBuilders.matchQuery(NAME_FIELD, name)
+        ).setRefresh(true);
 
-            BulkByScrollResponse response = client.execute(DeleteByQueryAction.INSTANCE, request).actionGet();
+        BulkByScrollResponse response = client.execute(DeleteByQueryAction.INSTANCE, request).actionGet();
 
-            long deletedDocs = response.getDeleted();
+        long deletedDocs = response.getDeleted();
 
-            if (deletedDocs == 0) {
-                throw new ApiTokenException("No token found with name " + name);
-            }
-        });
+        if (deletedDocs == 0) {
+            throw new ApiTokenException("No token found with name " + name);
+        }
     }
 
     public Map<String, ApiToken> getTokenMetadatas() {
-        return withSecurityContext(() -> {
-            try {
-                SearchRequest searchRequest = new SearchRequest(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX);
-                searchRequest.source(new SearchSourceBuilder());
+        try {
+            SearchRequest searchRequest = new SearchRequest(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX);
+            searchRequest.source(new SearchSourceBuilder());
 
-                SearchResponse response = client.search(searchRequest).actionGet();
+            SearchResponse response = client.search(searchRequest).actionGet();
 
-                Map<String, ApiToken> tokens = new HashMap<>();
-                for (SearchHit hit : response.getHits().getHits()) {
-                    try (
-                        XContentParser parser = XContentType.JSON.xContent()
-                            .createParser(
-                                NamedXContentRegistry.EMPTY,
-                                DeprecationHandler.THROW_UNSUPPORTED_OPERATION,
-                                hit.getSourceRef().streamInput()
-                            )
-                    ) {
+            Map<String, ApiToken> tokens = new HashMap<>();
+            for (SearchHit hit : response.getHits().getHits()) {
+                try (
+                    XContentParser parser = XContentType.JSON.xContent()
+                        .createParser(
+                            NamedXContentRegistry.EMPTY,
+                            DeprecationHandler.THROW_UNSUPPORTED_OPERATION,
+                            hit.getSourceRef().streamInput()
+                        )
+                ) {
 
-                        ApiToken token = ApiToken.fromXContent(parser);
-                        tokens.put(token.getName(), token);
-                    }
+                    ApiToken token = ApiToken.fromXContent(parser);
+                    tokens.put(token.getName(), token);
                 }
-                return tokens;
-            } catch (IOException e) {
-                throw new RuntimeException(e);
             }
-        });
+            return tokens;
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     public Boolean apiTokenIndexExists() {
         return clusterService.state().metadata().hasConcreteIndex(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX);
     }
 
-    private <T> T withSecurityContext(Supplier<T> operation) {
-        final var originalUserAndRemoteAddress = Utils.userAndRemoteAddressFrom(client.threadPool().getThreadContext());
-        try (final ThreadContext.StoredContext ctx = client.threadPool().getThreadContext().stashContext()) {
-            client.threadPool()
-                .getThreadContext()
-                .putTransient(ConfigConstants.OPENDISTRO_SECURITY_USER, originalUserAndRemoteAddress.getLeft());
-            return operation.get();
-        }
-    }
-
-    private void withSecurityContext(Runnable operation) {
-        final var originalUserAndRemoteAddress = Utils.userAndRemoteAddressFrom(client.threadPool().getThreadContext());
-        try (final ThreadContext.StoredContext ctx = client.threadPool().getThreadContext().stashContext()) {
-            client.threadPool()
-                .getThreadContext()
-                .putTransient(ConfigConstants.OPENDISTRO_SECURITY_USER, originalUserAndRemoteAddress.getLeft());
-            operation.run();
-        }
-    }
-
     public void createApiTokenIndexIfAbsent() {
         if (!apiTokenIndexExists()) {
-            final var originalUserAndRemoteAddress = Utils.userAndRemoteAddressFrom(client.threadPool().getThreadContext());
-            try (final ThreadContext.StoredContext ctx = client.threadPool().getThreadContext().stashContext()) {
-                client.threadPool()
-                    .getThreadContext()
-                    .putTransient(ConfigConstants.OPENDISTRO_SECURITY_USER, originalUserAndRemoteAddress.getLeft());
-                final Map<String, Object> indexSettings = ImmutableMap.of(
-                    "index.number_of_shards",
-                    1,
-                    "index.auto_expand_replicas",
-                    "0-all"
-                );
-                final CreateIndexRequest createIndexRequest = new CreateIndexRequest(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX).settings(
-                    indexSettings
-                );
-                client.admin().indices().create(createIndexRequest);
-            }
+            final Map<String, Object> indexSettings = ImmutableMap.of("index.number_of_shards", 1, "index.auto_expand_replicas", "0-all");
+            final CreateIndexRequest createIndexRequest = new CreateIndexRequest(ConfigConstants.OPENSEARCH_API_TOKENS_INDEX).settings(
+                indexSettings
+            );
+            client.admin().indices().create(createIndexRequest);
         }
     }
